backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Admin <admin@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.18.2"
fastapi = "^0.70.1"
python-multipart = "^0.0.5"
email-validator = "^1.0.5"
requests = "^2.23.0"
passlib = {extras = ["bcrypt"], version = "^1.7.2"}
tenacity = "^6.1.0"
pydantic = "^1.4"
emails = "^0.5.15"
raven = "^6.10.0"
gunicorn = "^20.0.4"
jinja2 = "^2.11.2"
psycopg2-binary = "^2.8.5"
alembic = "^1.4.2"
sqlalchemy = "^1.3.16"
python-jose = {extras = ["cryptography"], version = "^3.1.0"}
markupsafe = "2.0.1"
dicomweb-client = "^0.56.2"

[tool.poetry.dev-dependencies]
mypy = "^0.770"
black = "^22.6.0"
isort = "^4.3.21"
autoflake = "^1.3.1"
flake8 = "^3.7.9"
pytest = "^7.1.2"
sqlalchemy-stubs = "^0.3"
pytest-cov = "^2.8.1"
jupyter = "^1.0.0"
ipykernel = "6.2.0"
pytest-order = "^1.0.0"
rstr = "^3.0.0"
pandas = "^1.3.5"
watchgod = "^0.8.2"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
